requirements.txt
build
click
coverage
coveralls
daiquiri
datashader
hvplot
jupyter-book
matplotlib
msprime
panel
pre-commit
pytest
pytest-playwright
selenium
sphinx_issues
tskit
tszip
